+++ tests/skeleton/resources/views/partials/has_use.blade.php
@@ -1,3 +1,4 @@
 @php
 use My\Name\Space;
 @endphp
+{{ 'Please use App\Foo; thanks' }}